+++ CBV-with-CBN-embedding/js/nodes/recur.js
@@ -13,38 +13,38 @@ class Recur extends Expo {
 	}
 
 	rewrite(token, nextLink) {
-		if (nextLink.from == this.key) {
-			if (token.rewriteFlag == RewriteFlag.F_RECUR) {
-				token.rewriteFlag = RewriteFlag.EMPTY;
+		if (token.rewriteFlag == RewriteFlag.F_RECUR && nextLink.from == this.key) {
+			token.rewriteFlag = RewriteFlag.EMPTY;
 
-				var wrapper = BoxWrapper.create().addToGroup(this.group.group);
-				var newBox = this.group.copy().addToGroup(wrapper.box);
-				wrapper.auxs = wrapper.createPaxsOnTopOf(newBox.auxs);
-				
-				new Link(wrapper.prin.key, newBox.prin.key, "n", "s").addToGroup(wrapper);
-				
-				var oldGroup = this.group;
-				var oldBox = this.group.box;
-				this.group.moveOut();
-				var leftLink = this.findLinksInto("w")[0];
-				leftLink.changeTo(wrapper.prin.key, "s");
-				leftLink.fromPort = "n";
-				leftLink.reverse = false;
-				var inLink = this.findLinksInto("s")[0];
-				var outLink = this.findLinksOutOf("e")[0];
-				outLink.changeFrom(inLink.from, inLink.fromPort);
+			var wrapper = BoxWrapper.create().addToGroup(this.group.group);
+			var newBox = this.group.copy().addToGroup(wrapper.box);
+			wrapper.auxs = wrapper.createPaxsOnTopOf(newBox.auxs);
+			
+			new Link(wrapper.prin.key, newBox.prin.key, "n", "s").addToGroup(wrapper);
+			
+			var oldGroup = this.group;
+			var oldBox = this.group.box;
+			this.group.moveOut();
+			var leftLink = this.findLinksInto("w")[0];
+			leftLink.changeTo(wrapper.prin.key, "s");
+			leftLink.fromPort = "n";
+			leftLink.reverse = false;
+			var inLink = this.findLinksInto("s")[0];
+			var outLink = this.findLinksOutOf("e")[0];
+			outLink.changeFrom(inLink.from, inLink.fromPort);
 
-				Term.joinAuxs(wrapper.auxs, oldBox.auxs, wrapper.group);
-				
-				oldGroup.delete();
+			Term.joinAuxs(wrapper.auxs, oldBox.auxs, wrapper.group);
+			
+			oldGroup.delete();
 
-				token.rewrite = true;
-				return nextLink;
-			}
+			token.rewrite = true;
+			return nextLink;
 		}
 
-		token.rewrite = false;
-		return nextLink;
+		else if (token.rewriteFlag == RewriteFlag.EMPTY) {
+			token.rewrite = false;
+			return nextLink;
+		}
 	}
 
 	copy() {
