+++ CBV-with-CBV-embedding/js/nodes/contract.js
@@ -16,39 +16,36 @@ class Contract extends Expo {
 	}
 
 	rewrite(token, nextLink) {
-		if (nextLink.from == this.key) {
-			if (token.rewriteFlag == RewriteFlag.F_C) {
-				token.rewriteFlag = RewriteFlag.EMPTY;
+		if (token.rewriteFlag == RewriteFlag.F_C && nextLink.from == this.key) {
+			token.rewriteFlag = RewriteFlag.EMPTY;
 
-				if (this.findLinksInto(null).length == 1) {
+			if (this.findLinksInto(null).length == 1) {
+				token.boxStack.pop();
+				var inLink = this.findLinksInto(null)[0];
+				nextLink.changeFrom(inLink.from, inLink.fromPort);
+				this.delete();
+			}
+			else {
+				var i = token.boxStack.last();
+				var prev = this.graph.findNodeByKey(i.from);
+				if (prev instanceof Contract) {
 					token.boxStack.pop();
-					var inLink = this.findLinksInto(null)[0];
-					nextLink.changeFrom(inLink.from, inLink.fromPort);
-					this.delete();
-				}
-				else if (token.boxStack.length >= 2) {
-					var i = token.boxStack.last();
-					var prev = this.graph.findNodeByKey(i.from);
-					if (prev instanceof Contract) {
-						token.boxStack.pop();
-						for (let link of prev.findLinksInto(null)) {
-							link.changeTo(this.key, "s");
-						}
-						prev.delete();
+					for (let link of prev.findLinksInto(null)) {
+						link.changeTo(this.key, "s");
 					}
+					prev.delete();
 					token.rewriteFlag = RewriteFlag.F_C;
 				}
-				else if (token.boxStack.length == 1) {
-					
-				}
-
-				token.rewrite = true;
-				return nextLink;
 			}
+			
+			token.rewrite = true;
+			return nextLink;
 		}
-
-		token.rewrite = false;
-		return nextLink;
+		
+		else if (token.rewriteFlag == RewriteFlag.EMPTY) {
+			token.rewrite = false;
+			return nextLink;
+		}
 	}
 
 	copy() {
